Create GetPossibleAncestorRevisionIdsCallable
This is from getPossibleAncestorRevisionIDs and also extracts the
getPossibleAncestorRevisionIDsInQueue method from BasicDatastore
which wasn't used aside from in getPossibleAncestorRevisionIDs.

This callable was standalone so can be pulled out with very
few changes -- literally just the IDEA refactor works for this
one, with a class-comment as to the class's purpose.
